Rename Default[Method|Query]ExecutionInfo to Mutable~ExecutionInfo
