+++ src/main/java/io/r2dbc/proxy/callback/MutableMethodExecutionInfo.java
@@ -18,58 +18,47 @@
 package io.r2dbc.proxy.callback;
 
 import io.r2dbc.proxy.core.ConnectionInfo;
-import io.r2dbc.proxy.core.ExecutionType;
+import io.r2dbc.proxy.core.MethodExecutionInfo;
 import io.r2dbc.proxy.core.ProxyEventType;
-import io.r2dbc.proxy.core.QueryExecutionInfo;
-import io.r2dbc.proxy.core.QueryInfo;
 import io.r2dbc.proxy.util.Assert;
 
 import java.lang.reflect.Method;
 import java.time.Duration;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 /**
- * Default implementation of the {@link QueryExecutionInfo}.
+ * Default implementation of the {@link MethodExecutionInfo}.
  *
  * @author Tadaya Tsuyukubo
  */
-final class DefaultQueryExecutionInfo implements QueryExecutionInfo {
+final class MutableMethodExecutionInfo implements MethodExecutionInfo {
 
-    private ConnectionInfo connectionInfo;
+    private Object target;
 
     private Method method;
 
     private Object[] methodArgs;
 
-    private Throwable throwable;
+    private Object result;
 
-    private boolean isSuccess;
+    private Throwable thrown;
 
-    private int batchSize;  // num of Batch#add
-
-    private ExecutionType type;
-
-    private int bindingsSize;  // num of Statement#add
+    private ConnectionInfo connectionInfo;
 
     private Duration executeDuration = Duration.ZERO;
 
-    private String threadName = "";
+    private String threadName;
 
     private long threadId;
 
     private ProxyEventType proxyEventType;
 
-    private int currentResultCount;
-
-    private Object currentMappedResult;
-
-    private List<QueryInfo> queries = new ArrayList<>();
-
     private Map<String, Object> customValues = new HashMap<>();
 
+    public void setTarget(Object target) {
+        this.target = target;
+    }
 
     public void setMethod(Method method) {
         this.method = method;
@@ -79,40 +68,16 @@ final class DefaultQueryExecutionInfo implements QueryExecutionInfo {
         this.methodArgs = methodArgs;
     }
 
+    public void setResult(Object result) {
+        this.result = result;
+    }
+
+    public void setThrown(Throwable thrown) {
+        this.thrown = thrown;
+    }
+
     public void setConnectionInfo(ConnectionInfo connectionInfo) {
         this.connectionInfo = connectionInfo;
-    }
-
-    public void setThrowable(Throwable throwable) {
-        this.throwable = throwable;
-    }
-
-    public void setSuccess(boolean isSuccess) {
-        this.isSuccess = isSuccess;
-    }
-
-    public void setBatchSize(int batchSize) {
-        this.batchSize = batchSize;
-    }
-
-    public void setQueries(List<QueryInfo> queries) {
-        this.queries = queries;
-    }
-
-    public ExecutionType getType() {
-        return type;
-    }
-
-    public void setType(ExecutionType type) {
-        this.type = type;
-    }
-
-    public int getBindingsSize() {
-        return bindingsSize;
-    }
-
-    public void setBindingsSize(int bindingsSize) {
-        this.bindingsSize = bindingsSize;
     }
 
     public void setExecuteDuration(Duration executeDuration) {
@@ -131,14 +96,6 @@ final class DefaultQueryExecutionInfo implements QueryExecutionInfo {
         this.proxyEventType = proxyEventType;
     }
 
-    public void setCurrentResultCount(int currentResultCount) {
-        this.currentResultCount = currentResultCount;
-    }
-
-    public void setCurrentMappedResult(Object currentResult) {
-        this.currentMappedResult = currentResult;
-    }
-
     @Override
     public void addCustomValue(String key, Object value) {
         Assert.requireNonNull(key, "key must not be null");
@@ -155,6 +112,11 @@ final class DefaultQueryExecutionInfo implements QueryExecutionInfo {
     }
 
     @Override
+    public Object getTarget() {
+        return target;
+    }
+
+    @Override
     public Method getMethod() {
         return method;
     }
@@ -165,28 +127,18 @@ final class DefaultQueryExecutionInfo implements QueryExecutionInfo {
     }
 
     @Override
+    public Object getResult() {
+        return result;
+    }
+
+    @Override
+    public Throwable getThrown() {
+        return thrown;
+    }
+
+    @Override
     public ConnectionInfo getConnectionInfo() {
         return this.connectionInfo;
-    }
-
-    @Override
-    public Throwable getThrowable() {
-        return throwable;
-    }
-
-    @Override
-    public boolean isSuccess() {
-        return isSuccess;
-    }
-
-    @Override
-    public int getBatchSize() {
-        return batchSize;
-    }
-
-    @Override
-    public List<QueryInfo> getQueries() {
-        return this.queries;
     }
 
     @Override
@@ -207,16 +159,6 @@ final class DefaultQueryExecutionInfo implements QueryExecutionInfo {
     @Override
     public ProxyEventType getProxyEventType() {
         return proxyEventType;
-    }
-
-    @Override
-    public int getCurrentResultCount() {
-        return currentResultCount;
-    }
-
-    @Override
-    public Object getCurrentMappedResult() {
-        return currentMappedResult;
     }
 
 }
