+++ src/main/java/io/r2dbc/proxy/callback/ResultCallbackHandler.java
@@ -36,14 +36,13 @@ final class ResultCallbackHandler extends CallbackHandlerSupport {
 
     private final Result result;
 
-    private final DefaultQueryExecutionInfo queryExecutionInfo;
+    private final MutableQueryExecutionInfo queryExecutionInfo;
 
     /**
      * Callback handler logic for {@link Result}.
      *
-     * This constructor purposely uses {@link QueryExecutionInfo} interface for arguments instead of {@link DefaultQueryExecutionInfo} implementation.
-     * This way, creator of this callback handler ({@link ProxyFactory}) does not depend on {@link DefaultQueryExecutionInfo} implementation which
-     * is scoped to package private.
+     * This constructor purposely uses {@link QueryExecutionInfo} interface for arguments instead of {@link MutableQueryExecutionInfo} implementation.
+     * This way, creator of this callback handler ({@link ProxyFactory}) does not depend on {@link MutableQueryExecutionInfo} implementation.
      *
      * @param result             query result
      * @param queryExecutionInfo query execution info
@@ -51,13 +50,13 @@ final class ResultCallbackHandler extends CallbackHandlerSupport {
      * @throws IllegalArgumentException if {@code result} is {@code null}
      * @throws IllegalArgumentException if {@code queryExecutionInfo} is {@code null}
      * @throws IllegalArgumentException if {@code proxyConfig} is {@code null}
-     * @throws IllegalArgumentException if {@code queryExecutionInfo} is not an instance of {@link DefaultQueryExecutionInfo}
+     * @throws IllegalArgumentException if {@code queryExecutionInfo} is not an instance of {@link MutableQueryExecutionInfo}
      */
     public ResultCallbackHandler(Result result, QueryExecutionInfo queryExecutionInfo, ProxyConfig proxyConfig) {
         super(proxyConfig);
         this.result = Assert.requireNonNull(result, "result must not be null");
         Assert.requireNonNull(queryExecutionInfo, "queryExecutionInfo must not be null");
-        this.queryExecutionInfo = Assert.requireType(queryExecutionInfo, DefaultQueryExecutionInfo.class, "queryExecutionInfo must be DefaultQueryExecutionInfo");
+        this.queryExecutionInfo = Assert.requireType(queryExecutionInfo, MutableQueryExecutionInfo.class, "queryExecutionInfo must be MutableQueryExecutionInfo");
     }
 
     @Override
